requirements.txt
attrs==23.1.0
Automat==22.10.0
awscli==1.29.71
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.28.71
botocore==1.31.71
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
docutils==0.16
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
filelock==3.12.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kaitaistruct==0.10
lxml==4.9.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Pillow==10.0.0
Protego==0.3.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyDispatcher==2.0.7
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
rsa==4.7.2
rubicon-objc==0.4.6
s3transfer==0.7.0
Scrapy==2.11.0
selenium==4.11.2
selenium-wire==5.1.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==3.4.4
torpy==1.1.6
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
wsproto==1.2.0
zope.interface==6.0
zstandard==0.22.0
